+++ ios/RNBugly.m
@@ -1,6 +1,7 @@
 
 #import "RNBugly.h"
 #import <Bugly/Bugly.h>
+#import <React/RCTConvert.h>
 
 @implementation RNBugly
 
@@ -34,10 +35,48 @@ RCT_EXPORT_METHOD(setTag:(NSUInteger)tag)
     [Bugly setTag:tag];
 }
 
+RCT_EXPORT_METHOD(closeCrashReport)
+{
+    [Bugly closeCrashReport];
+}
+
+RCT_EXPORT_METHOD(getCurrentTag:(RCTPromiseResolveBlock)resolve
+                  reject:(RCTPromiseRejectBlock)reject)
+{
+    NSUInteger tag = [Bugly currentTag];
+    resolve([NSNumber numberWithUnsignedInteger:tag]);
+}
+
+RCT_EXPORT_METHOD(getUserData:(RCTPromiseResolveBlock)resolve
+                  reject:(RCTPromiseRejectBlock)reject)
+{
+    NSDictionary* dict = [Bugly allUserValues];
+    resolve(dict);
+}
+
+RCT_EXPORT_METHOD(getBuglyVersion:(RCTPromiseResolveBlock)resolve
+                  reject:(RCTPromiseRejectBlock)reject)
+{
+    resolve([Bugly sdkVersion]);
+}
+
 RCT_EXPORT_METHOD(putUserData:(NSString *)userKey:(NSString *)userValue)
 {
     [Bugly setUserValue:userValue forKey:userKey];
 }
 
+RCT_EXPORT_METHOD(postException:(NSDictionary *)params)
+{
+    NSInteger category = [RCTConvert NSInteger:[params objectForKey:@"category"]];
+    NSString *errorType = [RCTConvert NSString:[params objectForKey:@"errorType"]];
+    NSString *errorMsg = [RCTConvert NSString:[params objectForKey:@"errorMsg"]];
+    NSString *stack = [RCTConvert NSString:[params objectForKey:@"stack"]];
+    NSArray *stackTraceArray = [stack componentsSeparatedByString:@""];
+    NSDictionary *extraInfo = [RCTConvert NSDictionary:[params objectForKey:@"extraInfo"]];
+    if(extraInfo == nil) {
+        extraInfo = [NSMutableDictionary dictionary];
+    }
+    [Bugly reportExceptionWithCategory:category name:errorMsg reason:@" " callStack:stackTraceArray extraInfo:extraInfo terminateApp:NO];
+}
 @end
-  
+